[Query] Change GLib types to standard types.
 - Change type of the port from uint32 to uint16.
 - Change Glib types to standard types.

+++ gst/nnstreamer/tensor_query/tensor_query_serversink.h
@@ -43,7 +43,7 @@ typedef struct _GstTensorQueryServerSinkClass GstTensorQueryServerSinkClass;
 struct _GstTensorQueryServerSink
 {
   GstBaseSink element; /**< parent object */
-  guint32 port;
+  guint16 port;
   gchar *host;
   TensorQueryProtocol protocol;
   guint timeout;
